requirements.txt
async-generator==1.10
asyncio==3.4.3
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
distlib==0.3.6
exceptiongroup==1.0.4
filelock==3.8.0
fyers-apiv2==2.0.5
h11==0.14.0
idna==2.10
MouseInfo==0.1.3
numpy==1.23.5
outcome==1.2.0
pandas==1.5.2
platformdirs==2.5.4
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyotp==2.7.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
requests==2.25.1
selenium==4.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
TA-Lib @ file:///F:/TA_Lib-0.4.24-cp310-cp310-win_amd64.whl
tornado==6.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
virtualenv==20.17.0
websocket-client==1.2.1
websockets==8.1
wsproto==1.2.0
